feat: working search implementation with regex
